tests/requirements.txt
pytest==5.4.3
python-dateutil==2.8.1